requirements.txt
pydantic
pyyaml

geopandas
simplekml
rasterio
shapely
lxml
pyproj
